+++ lib/config/validation.js
@@ -1,5 +1,5 @@
 const options = require('./definitions').getOptions();
-const { isValidSemver } = require('../util/semver');
+const { isValidSemver } = require('../versioning/semver');
 const { resolveConfigPresets } = require('./presets');
 const {
   hasValidSchedule,
