+++ lib/datasource/github.js
@@ -1,5 +1,5 @@
 const ghGot = require('../platform/github/gh-got-wrapper');
-const { isPinnedVersion, semverSort } = require('../util/semver');
+const { isPinnedVersion, semverSort } = require('../versioning/semver');
 
 module.exports = {
   getRepoTags,
