+++ lib/manager/_helpers/node/package.js
@@ -3,7 +3,7 @@ const {
   getMinor,
   isGreaterThan,
   isPinnedVersion,
-} = require('../../../util/semver');
+} = require('../../../versioning/semver');
 const { getRepoTags, semverSort } = require('../../../datasource/github');
 
 async function getPackageUpdates(config) {
