+++ lib/manager/bazel/package.js
@@ -3,7 +3,7 @@ const {
   isGreaterThan,
   isStable,
   isPinnedVersion,
-} = require('../../util/semver');
+} = require('../../versioning/semver');
 const { getRepoTags, getRepoReleases } = require('../../datasource/github');
 
 module.exports = {
