+++ lib/manager/buildkite/package.js
@@ -4,7 +4,7 @@ const {
   semverSort,
   getMajor,
   getMinor,
-} = require('../../util/semver');
+} = require('../../versioning/semver');
 const { getRepoTags } = require('../../datasource/github');
 
 module.exports = {
