+++ lib/manager/docker/package.js
@@ -1,4 +1,4 @@
-const { getMajor, isValidSemver } = require('../../util/semver');
+const { getMajor, isValidSemver } = require('../../versioning/semver');
 const dockerApi = require('../../datasource/docker');
 const compareVersions = require('compare-versions');
