+++ lib/manager/npm/lookup.js
@@ -1,5 +1,5 @@
 // const datasource = require('../../datasource/npm');
-const versioning = require('../../util/semver.js');
+const versioning = require('../../versioning/semver.js');
 const moment = require('moment');
 
 const {
