+++ lib/manager/npm/package.js
@@ -1,8 +1,8 @@
 const npmApi = require('../../datasource/npm');
 const lookup = require('./lookup');
-const { isValidSemver } = require('../../util/semver');
+const { isValidSemver } = require('../../versioning/semver');
 const nodeManager = require('../_helpers/node/package');
-const { parseRange } = require('../../util/semver');
+const { parseRange } = require('../../versioning/semver');
 
 module.exports = {
   getRangeStrategy,
