+++ lib/manager/pip_requirements/package.js
@@ -4,7 +4,7 @@ const {
   semverSort,
   isPinnedVersion,
   getMajor,
-} = require('../../util/semver');
+} = require('../../versioning/semver');
 
 module.exports = {
   getPackageUpdates,
