+++ lib/manager/travis/package.js
@@ -1,6 +1,9 @@
 const { isEqual } = require('lodash');
 const { getRepoReleases } = require('../../datasource/github');
-const { isPinnedVersion, maxSatisfyingVersion } = require('../../util/semver');
+const {
+  isPinnedVersion,
+  maxSatisfyingVersion,
+} = require('../../versioning/semver');
 
 module.exports = {
   getPackageUpdates,
