+++ lib/util/package-rules.js
@@ -1,6 +1,6 @@
 const minimatch = require('minimatch');
 
-const { intersectsSemver } = require('./semver');
+const { intersectsSemver } = require('../versioning/semver');
 const { mergeChildConfig } = require('../config');
 
 module.exports = {
