+++ lib/workers/pr/changelog/manager-npm.js
@@ -1,5 +1,5 @@
 const npmRegistry = require('../../../datasource/npm');
-const { semverSort } = require('../../../util/semver');
+const { semverSort } = require('../../../versioning/semver');
 
 module.exports = {
   getPackage,
