+++ lib/workers/pr/changelog/manager-pip.js
@@ -1,5 +1,5 @@
 const got = require('got');
-const { semverSort, isPinnedVersion } = require('../../../util/semver');
+const { semverSort, isPinnedVersion } = require('../../../versioning/semver');
 
 module.exports = {
   getPackage,
