+++ lib/workers/pr/changelog/source-github.js
@@ -1,4 +1,7 @@
-const { matchesSemver, isPinnedVersion } = require('../../../util/semver');
+const {
+  matchesSemver,
+  isPinnedVersion,
+} = require('../../../versioning/semver');
 const ghGot = require('../../../platform/github/gh-got-wrapper');
 
 module.exports = {
